PythonAILib/python_ai_lib/dev/autogen_sample3_two-agent-chat.py
import os
from typing import Annotated
from dotenv import load_dotenv
import autogen  # type: ignore
import tempfile

from autogen import ConversableAgent
from autogen.coding import LocalCommandLineCodeExecutor # type: ignore

# Create a temporary directory to store the code files.
temp_dir = tempfile.TemporaryDirectory()

# Create a local command line code executor.
executor = LocalCommandLineCodeExecutor(
    timeout=10,  # Timeout for each code execution in seconds.
    work_dir=temp_dir.name,  # Use the temporary directory to store the code files.
)
load_dotenv()

llm_config = {
    "config_list": [{
        "model": "gpt-4o-mini", "api_key": os.environ["OPENAI_API_KEY"], }],
}

student_agent = ConversableAgent(
    name="Student_Agent",
    system_message="You are a student willing to learn.",
    llm_config={"config_list": [{"model": "gpt-4", "api_key": os.environ["OPENAI_API_KEY"]}]},
)
teacher_agent = ConversableAgent(
    name="Teacher_Agent",
    system_message="You are a math teacher.",
    llm_config={"config_list": [{"model": "gpt-4", "api_key": os.environ["OPENAI_API_KEY"]}]},
)

chat_result = student_agent.initiate_chat(
    teacher_agent,
    message="What is triangle inequality?",
    summary_method="reflection_with_llm",
    max_turns=2,
)

print(chat_result.cost)