requirements.txt
pillow
python-docx
python-dotenv
python-pptx
tiktoken
beautifulsoup4
chromadb
openai==1.52.1
langchain==0.3.4
langchain-openai==0.2.3
langchain-community==0.3.3
langchain-postgres==0.0.12
magika
openpyxl
markdown
pdfminer.six
chardet
SQLAlchemy
langchain_postgres
psycopg-binary
autogenstudio==0.1.5
autogen-agentchat~=0.2
wikipedia
flaml[automl]
selenium
webdriver-manager
duckduckgo-search